% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proc.content.R
\name{Proc.content}
\alias{Proc.content}
\title{Content table}
\usage{
Proc.content(x)
}
\arguments{
\item{x}{A data.frame() object}
}
\value{
returns a content table
}
\description{
Content table
}
\examples{
Proc.content(df)
}
