% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proc.reg.anova.R
\name{Proc.reg.anova}
\alias{Proc.reg.anova}
\title{Summary table of Anova test}
\usage{
Proc.reg.anova(lm)
}
\arguments{
\item{lm}{A lm() object}
}
\value{
returns a summary table of Anova test
}
\description{
Summary table of Anova test
}
\examples{
\dontrun{Proc.reg.anova(lm)}
}
