% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proc.reg.confint.R
\name{Proc.reg.confint}
\alias{Proc.reg.confint}
\title{Summary table of a linear model}
\usage{
Proc.reg.confint(lm)
}
\arguments{
\item{lm}{A lm() object}
}
\value{
returns a summary table of the confidence intervals from the paraters calculated on a lineal model
}
\description{
Summary table of a linear model
}
\examples{
\dontrun{Proc.reg.confint(lm)}
}
