% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proc.reg.durbin.R
\name{Proc.reg.durbin}
\alias{Proc.reg.durbin}
\title{Summary table of Durbin test}
\usage{
Proc.reg.durbin(lm)
}
\arguments{
\item{lm}{A lm() object}
}
\value{
returns a summary table of Durbin test
}
\description{
Summary table of Durbin test
}
\examples{
Proc.reg.durbin(lm)
}
