% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proc.reg.pagan.R
\name{Proc.reg.pagan}
\alias{Proc.reg.pagan}
\title{Summary table of Breusch-Pagan test}
\usage{
Proc.reg.pagan(lm)
}
\arguments{
\item{lm}{A lm() object}
}
\value{
returns a summary table of Breusch-Pagan test
}
\description{
Summary table of Breusch-Pagan test
}
\examples{
\dontrun{Proc.reg.pagan(lm)}
}
