% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proc.reg.reset.R
\name{Proc.reg.reset}
\alias{Proc.reg.reset}
\title{Summary table of the Reset test}
\usage{
Proc.reg.reset(lm)
}
\arguments{
\item{lm}{A lm() object}
}
\value{
returns a summary table of the Reset test
}
\description{
Summary table of the Reset test
}
\examples{
Proc.reg.reset(lm)
}
