% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proc.reg.table.R
\name{Proc.reg.table}
\alias{Proc.reg.table}
\title{Summary table of a linear model}
\usage{
Proc.reg.table(lm)
}
\arguments{
\item{lm}{A lm() object}
}
\value{
returns a summary table of a linear model
}
\description{
Summary table of a linear model
}
\examples{
Proc.reg.table(lm)
}
