% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proc.reg.vifs.R
\name{Proc.reg.vifs}
\alias{Proc.reg.vifs}
\title{Plot of Vifs from a linear model}
\usage{
Proc.reg.vifs(lm)
}
\arguments{
\item{lm}{A lm() object}
}
\value{
Returns a graphic of Vifs from a linear model
}
\description{
Plot of Vifs from a linear model
}
\examples{
Proc.reg.vifs(lm)
}
