% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proc.reg.white.R
\name{Proc.reg.white}
\alias{Proc.reg.white}
\title{Summary table of the White test}
\usage{
Proc.reg.white(lm)
}
\arguments{
\item{lm}{A lm() object}
}
\value{
returns a summary table of the White test
}
\description{
Summary table of the White test
}
\examples{
Proc.reg.white(lm)
}
