% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{package}
\name{Proc.sas}
\alias{Proc.sas}
\alias{Proc.sas-package}
\title{Auxiliar functions for regression analysis}
\description{
Auxiliar functions for data and regression analysis, inspired by SAS procedures.
}
\author{
Ignasi Pascual Vilaseca
}
