% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proc.summary.R
\name{Proc.summary}
\alias{Proc.summary}
\title{Summary table of a dataset}
\usage{
Proc.summary(x)
}
\arguments{
\item{lm}{A data.frame() object}
}
\value{
returns a numeric and categoric summary table a dataset
}
\description{
Summary table of a dataset
}
\examples{
\dontrun{Proc.summary(df)}
}
